+++ packages/device_info_plus/example/integration_test/device_info_plus_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart=2.9
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:device_info_plus_tizen/device_info_plus_tizen.dart';
 import 'package:integration_test/integration_test.dart';
@@ -11,7 +9,7 @@ import 'package:integration_test/integration_test.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  TizenDeviceInfo tizenInfo;
+  late TizenDeviceInfo tizenInfo;
 
   setUpAll(() async {
     final deviceInfoPlugin = DeviceInfoPluginTizen();