+++ packages/video_player_videohole/example/lib/main.dart
@@ -371,8 +371,7 @@ class _DrmRemoteVideoState2 extends State<_DrmRemoteVideo2> {
 }
 
 class _ControlsOverlay extends StatelessWidget {
-  const _ControlsOverlay({Key? key, required this.controller})
-      : super(key: key);
+  const _ControlsOverlay({required this.controller});
 
   static const List<Duration> _exampleCaptionOffsets = <Duration>[
     Duration(seconds: -10),