+++ packages/device_info_plus/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Increase the minimum Flutter version to 3.3.
+
 ## 1.1.0
 
 * Remove dependency on `device_info_plus_platform_interface`.